init molecule scen. for kick-starting specifc ver of SD
There is some missing logic here, for instance, while the molecule
scenario will fire up an existing SD code-base - the packages are pulled
from production. This is fine if you are trying to match the last public
release but more problematic if you need to rewind even farther. Since
we don't have that requirement right now I'm punting that problem into
the future.